commit all data base

+++ ERP.Data/ModelsERP/product.cs
@@ -35,7 +35,7 @@ namespace ERP.Data.ModelsERP
         [Column(TypeName = "text")]
         public string pu_description { get; set; }
 
-        public byte? pu_unit { get; set; }
+        public int? pu_unit { get; set; }
 
         public int? product_category_id { get; set; }
